Time theorem
Time theorem
the constant of light speed and time may be determined from the pull between local space time distortion and universal space time stretch. 

Imagine light like a car and distance it traveled determined the time it took it to get from point a to b. Imagine curves in the road increasing the distance that must be driven, which, increases the time of the trip. 
Next, Imagine the expansion of the universe as a current like a tide or river. The expansion is shown to accelerate. Ok, this means that the distance increases from point a to b, and again light is slowed not by its speed but by its distance. 

Then imagine black holes and large mass bending and pulling space time. These change the local curvature.

Finally, notice that distant galaxies are not pulled into little pieces but simply pulled away from one another. Accirding to some research this distance is accelerating. Their mass seems to resist being shredded. They act as a unit.

To put it all together, it appears light speeds is dictated by the path light must take which determines its time from point a to b. This is determined by the competition of forces between local such as mass of the galaxy, star , black hole etc with its distortion and universal stretch/ expansion. 
In addition, it may be and seems in my mind that time is also a product of these tensions. Which would mean it is expressed differently from one galaxy to the next and allows light speed and time to be a constant in that environment dictated by the tensions of local pull vs universal stretch and specific and constant based on the local ( galaxy, other large mass or space time distortion) pull.
philolisophically, time is the change of things. If the change grows due to elongation of space time equaling or beyond what the fastest rate of action ( light speed) then time stops.

So its not the universal stretch / acceleration but more of the local pull I purpose that determines the constant of time and light speed.

c_local ∝ √(local gravitational compression ÷ cosmological stretch)

or 

 C eff = C Tnet/Tlocal = C ( 1-( s universe / T local ))





 I was helped with the equations by chat got and grok ai